Revert "Merge pull request #138 from lemenkov/python3"
This reverts commit 68dfbebe0b8ac9310f29737912e0b18ee8ff6fa0, reversing
changes made to 8ce2d8c21e38ae063774c891ce1b315aa28d9ae4.

+++ foobnix/util/time_utils.py
@@ -4,6 +4,7 @@ Created on Feb 26, 2010
 @author: ivan
 '''
 
+
 def size2text(size):
     if size > 1024 * 1024 * 1024:
         return "%.2f Gb" % (size / (1024 * 1024 * 1024.0))
@@ -16,13 +17,13 @@ def size2text(size):
 def convert_seconds_to_text(time_sec):
         time_sec = int(time_sec)
 
-        hours = int(time_sec / (60 * 60))
+        hours = time_sec / (60 * 60)
         time_sec = time_sec - (hours * 60 * 60)
 
-        mins = int(time_sec / 60)
+        mins = time_sec / 60
         time_sec = time_sec - (mins * 60)
 
-        secs = int(time_sec)
+        secs = time_sec
         if hours > 0:
             return '%(hours)d:%(mins)02d:%(secs)02d' % {'hours' : hours, 'mins': mins, 'secs': secs }
         else:
